Ignore removed metadata from packaging org
This is a temporary workaround for the deployment failures caused by the
`gem0` packaging org not having the 'Delete Managed Components' perm
enabled. Since the metadata has already been removed from the final
packaging namespace (`gem`) this is a low-risk change. Per discussions
with the team this is the preferred solution over another namespace
switch.